src/pyproject.toml
[tool.poetry]
name = "cv-analizador"
version = "0.1.0"
description = ""
authors = ["Davidson Delo  <126816671+alfredobumba@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.149.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
tinydb = "^4.8.2"
pydantic = "^2.9.2"
sqlalchemy = "^2.0.36"
mysql-connector-python = "^9.1.0"
streamlit = "^1.39.0"
streamlit-aggrid = "^1.0.5"
pandas = "^2.2.3"
langchain-groq = "^0.2.0"
python-dotenv = "^1.0.1"
pymupdf = "^1.24.12"
matplotlib = "^3.9.2"
plotly = "^5.24.1"
chardet = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
